+++ libs/libmif/src/io/mif_openfpga_format.h
@@ -1,7 +1,9 @@
 #ifndef MIF_OPENFPGA_FORMAT_H
 #define MIF_OPENFPGA_FORMAT_H
 
+#include <cstddef>
 #include <ostream>
+#include <string>
 
 #include "mif_storage.h"
 
@@ -18,6 +20,17 @@
 namespace openfpga {
 
 void serialize_openfpga_mif(const MifStorage& storage, std::ostream& os);
+
+/* Parse one raw MIF input line (comments stripped inside). */
+int parse_mif_line(const std::string& file_path, size_t line_no,
+                   const std::string& raw_line, MifStorage& mif_storage,
+                   bool& has_current_segment, MifSegmentId& current_segment_id,
+                   size_t& total_words);
+
+/* Validate parse results and drop a trailing empty segment if needed. */
+int finalize_openfpga_mif_parse(MifStorage& mif_storage, bool has_current_segment,
+                                const MifSegmentId& current_segment_id,
+                                size_t total_words, const std::string& file_path);
 
 } /* namespace openfpga */
